// mobile/src/types/Common.ts
export type UserType = "farmer" | "institution";
export type ProposalStatus = "pending" | "accepted" | "rejected" | "canceled";
export type CallStatus = "active" | "closed" | "canceled";
export type Months =
  | "Jan"
  | "Fev"
  | "Mar"
  | "Abr"
  | "Mai"
  | "Jun"
  | "Jul"
  | "Ago"
  | "Set"
  | "Out"
  | "Nov"
  | "Dez";
export type Frequency = "weekly" | "biweekly" | "monthly" | "bimonthly";
export type Unities = "kilograms" | "units" | "boxes" | "liters"
export type Categories = "vegetables" | "fruits" | "legumes" | "roots and tubers" | "grains and cereals" | "dairy" | string